+++ R/darwinator-package.R
@@ -1,0 +1,9 @@
+#' darwinator
+#'
+#' This package provides tools for working with sampling event
+#' datasets in DwcA (Darwin Core Archive) format
+#'
+#' @name darwinator-package
+#' @aliases darwinator
+#' @keywords package
+NULL