+++ R/zzz.R
@@ -1,0 +1,14 @@
+.onAttach <- function(lib, pkg){
+
+  # echo "darwinator" | toilet -f smblock
+
+  welcome <-
+"
+ _|  _. ._     o ._   _. _|_  _  ._
+(_| (_| | \\/\\/ | | | (_|  |_ (_) |
+
+"
+  packageStartupMessage(welcome)
+
+}
+