+++ NEWS.md
@@ -1,7 +1,17 @@
+NEWS
+====
+
+For more fine-grained list of changes or to report a bug, consult 
+
+* [The issues log](https://github.com/GBIF-Europe/darwinator/issues)
+* [The commit log](https://github.com/GBIF-Europe/darwinator/commits/master)
+
+## darwinator 0.1.2
+
+* Added dwc_terms dataset with recommended and simple Darwin Core terms
+
 # darwinator 0.1.1
 
 * Fixed some download issues for Norwegian sampling event datasets
 * Added a `NEWS.md` file to track changes to the package.
-
-